+++ aliforman/app/index.html
@@ -6,9 +6,12 @@
   <link rel="stylesheet" type="text/css" href="/base.css">
 </head>
 <body data-ng-app="demoApp">
-  <todo-list
-    data-list-title="TO DO LIST"
-    data-todos="todos"></todo-list>
+  <h2>Angular Directives Intrduction</h2>
+  <main>
+    <todo-list
+      data-list-title="TO DO LIST"
+      data-todos="todos"></todo-list>
+  </main>
   <script src="bundle.js"></script>
 </body>
 </html>
